fix: ContentBlocking fluent setters + StorageController Long flags
GeckoRuntimeHolder:
- ContentBlocking.Settings uses fluent setters that return 'this' — Kotlin
  does not map these to property assignments. Switch to method-call style:
  .antiTracking(...) .cookieBehavior(...) .enhancedTrackingProtectionLevel(...)
  .safeBrowsing(...) .strictSocialTrackingProtection(false)

NightPOSApplication:
- StorageController.ClearFlags constants are Long in GeckoView 142
  (clearData() signature is clearData(@StorageClearFlags long flags)).
  Changed releaseGeckoCaches(flags: Int) → releaseGeckoCaches(flags: Long).

+++ app/src/main/java/com/nightpos/app/NightPOSApplication.kt
@@ -133,7 +133,7 @@ class NightPOSApplication : Application() {
      * Only NETWORK_CACHE and IMAGE_CACHE are used — never COOKIES or
      * DOM_STORAGES, which would destroy the Odoo session and IndexedDB data.
      */
-    private fun releaseGeckoCaches(flags: Int) {
+    private fun releaseGeckoCaches(flags: Long) {
         appScope.launch {
             runCatching {
                 GeckoRuntimeHolder.runtime.storageController.clearData(flags)
